+++ CHANGELOG.md
@@ -22,6 +22,10 @@ Bug-fixes within the same version aren't needed
 
 * upgrade to the latest jest version (24.7.x) - connectdotz
 
+* `Settings.settings` no longer contains the default configuration but starts with `undefined` - stephtr
+
+  Before accessing this property, one has to run `Settings.getConfig`.
+
 -->
 
 ### 25.0.0